+++ packages/winnow/test/project.js
@@ -2,6 +2,8 @@
 const test = require('tape')
 const winnow = require('../src')
 const features = require('./fixtures/snow.json').features
+const polygonFeatures = require('./fixtures/polygon.json').features
+const multiPolyFeatures = require('./fixtures/multipolygon.json').features
 
 test('Project to Web Mercator using 3857', t => {
   t.plan(2)
@@ -27,10 +29,54 @@ test('Project to Web Mercator using 3857 and translating to esri', t => {
   t.equal(results[0].geometry.y, 4857924.005275469)
 })
 
+test('Project a polygon to Web Mercator using 3857 and translating to esri', t => {
+  t.plan(2)
+  const options = {
+    projection: 3857,
+    limit: 1,
+    toEsri: true
+  }
+  const results = winnow.query(polygonFeatures, options)
+
+  t.equal(results.length, 1)
+  t.deepEqual(results[0].geometry.rings, [
+    [
+      [-13247454.246160466, 6496535.908013698],
+      [-11408073.597505985, 6985732.8890388245],
+      [-9216471.12251341, 4813698.293287256],
+      [-10214432.963804673, 3874440.0897190133],
+      [-12934368.178304385, 3424378.867175897],
+      [-13247454.246160466, 6496535.908013698]
+    ]
+  ])
+})
+
+test('Project a multi-polygon to Web Mercator using 3857 and translating to esri', t => {
+  t.plan(3)
+  const options = {
+    projection: 3857,
+    limit: 1,
+    toEsri: true
+  }
+  const results = winnow.query(multiPolyFeatures, options)
+
+  t.equal(results.length, 1)
+  const expected = [
+    [-13247454.246160466, 6496535.908013698],
+    [-11408073.597505985, 6985732.8890388245],
+    [-9216471.12251341, 4813698.293287256],
+    [-10214432.963804673, 3874440.0897190133],
+    [-12934368.178304385, 3424378.867175897],
+    [-13247454.246160466, 6496535.908013698]
+  ]
+  t.deepEqual(results[0].geometry.rings[0][0], expected[0])
+  t.deepEqual(results[0].geometry.rings[1][0], expected[0])
+})
+
 test('Project to Web Mercator using 3857 with an esri style outSR', t => {
   t.plan(1)
   const options = {
-    outSR: {latestWkid: 3857},
+    outSR: { latestWkid: 3857 },
     limit: 1
   }
   const results = winnow.query(features, options)